# PERF: StringArray construction