@TomAugspurger 2020-09-14
I'd prefer not to have a public argument that allows people to create an invalid object.

> The check is not needed in cases where we explicitly already have converted.

Can you say a bit more / give an example where we're doing the conversion / validation twice? Keep in mind that even if we know that we have all strings, we want to normalize all the NA values to `pd.NA`.