@jorisvandenbossche 2020-09-14
> Can you say a bit more / give an example where we're doing the conversion / validation twice? 

`_from_sequence` converts the input to an object array of all strings (or nans), and then `__init__` validates that the input are all strings (which we know is the case, as `_from_sequence` already ensured the conversion)