@TomAugspurger 2020-09-14
Right, Thanks.

Thoughts on refactoring `_from_sequence` to use `object.__new__(StringArray)` and then manually initializing by setting `_ndarray` and `_dtype` (and skipping the conversion / checks we don't need?) My preference would be to keep this out of the public API given the relative immaturity of StringArray.